refactored anecdote list to separate module
