+++ part6/redux-anecdotes/src/App.jsx
@@ -1,29 +1,10 @@
-import { useSelector, useDispatch } from "react-redux";
-import { createAnecdote, upvote } from "./reducers/anecdoteReducer";
 import AnecdoteForm from "./components/AncedoteForm";
+import AnecdoteList from "./components/AnecdoteList";
 
 const App = () => {
-  const anecdotes = useSelector((state) => state);
-  const dispatch = useDispatch();
-
-  const sortedAnecdotes = [...anecdotes].sort((a, b) => b.votes - a.votes);
-
-  const vote = (id) => {
-    dispatch(upvote(id));
-  };
-
   return (
     <div>
-      <h2>Anecdotes</h2>
-      {sortedAnecdotes.map((anecdote) => (
-        <div key={anecdote.id}>
-          <div>{anecdote.content}</div>
-          <div>
-            has {anecdote.votes}
-            <button onClick={() => vote(anecdote.id)}>vote</button>
-          </div>
-        </div>
-      ))}
+      <AnecdoteList />
       <AnecdoteForm />
     </div>
   );
