+++ part6/redux-anecdotes/src/components/AncedoteForm.jsx
@@ -3,6 +3,7 @@ import { createAnecdote } from "../reducers/anecdoteReducer";
 
 const AnecdoteForm = () => {
   const dispatch = useDispatch();
+
   const addAnecdote = (event) => {
     event.preventDefault();
     const contentTarget = event.target.content;
