+++ part6/redux-anecdotes/src/components/AnecdoteList.jsx
@@ -1,0 +1,30 @@
+import { useSelector, useDispatch } from "react-redux";
+import { upvote } from "../reducers/anecdoteReducer";
+
+const AnecdoteList = () => {
+  const anecdotes = useSelector((state) => state);
+  const dispatch = useDispatch();
+
+  const sortedAnecdotes = [...anecdotes].sort((a, b) => b.votes - a.votes);
+
+  const vote = (id) => {
+    dispatch(upvote(id));
+  };
+
+  return (
+    <>
+      <h2>Anecdotes</h2>
+      {sortedAnecdotes.map((anecdote) => (
+        <div key={anecdote.id}>
+          <div>{anecdote.content}</div>
+          <div>
+            has {anecdote.votes}
+            <button onClick={() => vote(anecdote.id)}>vote</button>
+          </div>
+        </div>
+      ))}
+    </>
+  );
+};
+
+export default AnecdoteList;
